Refactor the operation manager into sub job classes.
